# openapi.yaml
openapi: 3.1.0
info:
  title: PUMEX RAG API
  description: Query the Pumex knowledge base using semantic search
  version: 1.0.0
servers:
  - url: https://pumex-rag-production.up.railway.app
    description: Production server
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: API key for authentication
paths:
  /api/query:
    post:
      operationId: queryKnowledgeBase
      summary: Search the Pumex knowledge base
      description: Performs semantic search across the Pumex knowledge base. Returns relevant document chunks ranked by truth status (canonical > current > context > archive). Drafts are excluded by default unless explicitly requested.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - question
              properties:
                question:
                  type: string
                  description: The user's question or search query
                limit:
                  type: integer
                  description: Maximum number of results to return
                  default: 5
                  minimum: 1
                  maximum: 20
                include_drafts:
                  type: boolean
                  description: Include documents from drafts folder
                  default: false
      responses:
        '200':
          description: Successful query
          content:
            application/json:
              schema:
                type: object
                properties:
                  question:
                    type: string
                  count:
                    type: integer
                  results:
                    type: array
                    items:
                      type: object
                      properties:
                        title:
                          type: string
                        path:
                          type: string
                        status:
                          type: string
                        content:
                          type: string
                        similarity:
                          type: number
                        weighted_score:
                          type: number
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
  /api/learn:
    post:
      operationId: saveLearnedKnowledge
      summary: Save knowledge learned from conversations
      description: Allows ChatGPT to save new knowledge learned from conversations into the knowledge base. The content will be chunked, embedded, and made searchable.
      security:
        - ApiKeyAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - title
                - content
              properties:
                title:
                  type: string
                  description: Title of the learned knowledge
                  example: "New Protocol Feature"
                content:
                  type: string
                  description: The knowledge content to save (markdown supported)
                  example: "# New Protocol Feature\n\nThis feature was added in version 2.0..."
                source_question:
                  type: string
                  description: The original question that led to this knowledge
                source_conversation:
                  type: string
                  description: Context from the conversation where this was learned
      responses:
        '200':
          description: Knowledge saved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                  learned_id:
                    type: string
                  chunks_saved:
                    type: integer
                  message:
                    type: string
        '400':
          description: Bad request
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
        '500':
          description: Server error
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                  message:
                    type: string